Fixed bug in sentenceRanking

+++ StareProject/DataStructs.h
@@ -20,9 +20,9 @@ struct sentenceRanking
 	int sentenceID;		  // The selected sentence in the document that is being compared.
 	int foundDocumentID;  // the document id on the server that matches the new source sentence.
 	int foundSentenceID;  // The sentence id on the server that most closely matches the source sentence.
-	float certainty;	  // The certainty that a certain sentence is the particular style listed.
+	double certainty;	  // The certainty that a certain sentence is the particular style listed.
 
-	sentenceRanking(int SentenceID, int FoundDocumentID, int FoundSentenceID, float Certainty)
+	sentenceRanking(int SentenceID, int FoundDocumentID, int FoundSentenceID, double Certainty)
 	{
 		sentenceID = SentenceID;
 		foundDocumentID = FoundDocumentID;
